% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{mnsl_hues}
\alias{mnsl_hues}
\title{Munsell hues}
\usage{
mnsl_hues()
}
\value{
a character vector containing the hue values.
}
\description{
Returns a character vector of the Munsell hues in hue order starting at 2.5R and excluding grey ("N").
}
\examples{
mnsl_hues()
}
