\name{munsell.map}
\alias{munsell.map}
\docType{data}
\title{Provides link between Munsell colours hex and LUV}
\description{
Lookup table for converting Munsell colours to hex and LUV.
}
\usage{munsell.map}
\format{A dataframe containing columns hue, chroma, value, hex, name, L, U, V for 2740 Munsell colours.}
\source{Munsell Renotation Data, RIT Munsell Color Science Laboratory.  \url{http://www.cis.rit.edu/mcsl/online/munsell.php}}
\references{

}
\keyword{datasets}
