\name{plot.hex}
\alias{plot.hex}
\title{Plot hex colours...}
\usage{plot.hex(hex.colours, back.col=white)}
\description{Plot hex colours}
\details{Quick way to look at a set of hex colours.}
\value{A ggplot object}
\arguments{\item{hex.colour}{character vector specifying colours in hex form}
\item{back.col}{specification of background colour of display}}
\examples{plot.hex("#000000")
plot.hex(c("#000000","#FFFFFF"))}
