\name{text_colour}
\alias{text_colour}
\title{Get text colour...}
\usage{text_colour(cols)}
\description{Get text colour}
\details{Get the appropriate text colour for writing on a munsell colour.}
\value{a vector of "black" or "white"}
\keyword{internal}
\arguments{\item{a}{character vector of munsell colours}}
