% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.r
\name{theme_munsell}
\alias{theme_munsell}
\title{Default munsell plot theme}
\usage{
theme_munsell(bg.col = "white")
}
\arguments{
\item{bg.col}{takes colour to use as background colour}
}
\description{
Removes unnecessary clutter in plots
}
\keyword{internal}

