% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check_mnsl}
\alias{check_mnsl}
\title{Checks for valid Munsell colours}
\usage{
check_mnsl(col)
}
\arguments{
\item{col}{a character vector representing Munsell colours.}
}
\value{
a character vector containing the input colours.  If any colours
were outside the gamut they will be represented by NA.
}
\description{
Checks user supplied munsell specification for validity.  
I.e. colour is of form "h v/c" and h,  v and c take valid values.
}
\examples{
check_mnsl(c("5R 5/8","2.5R 9/28"))
}
\keyword{internal}
