% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter.r
\name{darker}
\alias{darker}
\title{Make a munsell colour darker}
\usage{
darker(col, steps = 1)
}
\arguments{
\item{col}{character vector of Munsell colours}

\item{steps}{number of steps to take in decreasing value}
}
\value{
character vector of Munsell colours
}
\description{
Decreases the value of the Munsell colour by 1.
}
\examples{
darker("5PB 3/4")
cols <- c("5PB 3/4", "5Y 7/8")
p <- plot_mnsl(c(cols, darker(cols), darker(cols, 2)))
p + ggplot2::facet_wrap(~ names, ncol = 2)
}

