\name{plot_common}
\alias{plot_common}
\title{Default display settings for plots of rectangular format...}
\usage{plot_common(bg.col)}

\description{
  Default display settings for plots of rectangular format
}
\keyword{internal}
\arguments{
  \item{bg.col}{takes colour to use as background colour}
}
