\name{text_colour}
\alias{text_colour}
\title{Get text colour}
\usage{
  text_colour(cols)
}
\arguments{
  \item{a}{character vector of munsell colours}
}
\value{
  a vector of "black" or "white"
}
\description{
  Get the appropriate text colour for writing on a munsell
  colour.
}
\keyword{internal}

